--- src/main/kotlin/com/sparta/reviewus/auth/exception/AuthExceptionHandler.kt
@@ -1,31 +1,0 @@
-package com.sparta.reviewus.auth.exception
-
-import com.sparta.reviewus.domain.exception.dto.ErrorResponse
-import org.springframework.http.HttpStatus
-import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.ExceptionHandler
-import org.springframework.web.bind.annotation.RestControllerAdvice
-
-@RestControllerAdvice
-class AuthExceptionHandler {
-
-    @ExceptionHandler(ExpiredJwtException::class)
-    fun handleExpiredJwtException(e: ExpiredJwtException): ResponseEntity<ErrorResponse> {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-            .body(ErrorResponse(e.message))
-    }
-
-    @ExceptionHandler(AuthenticationException::class)
-    fun handleAuthenticationException(e: AuthenticationException): ResponseEntity<ErrorResponse> {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-            .body(ErrorResponse(e.message))
-    }
-
-    @ExceptionHandler(BadCredentialsException::class)
-    fun handleBadCredentialsException(e: BadCredentialsException): ResponseEntity<ErrorResponse> {
-        return ResponseEntity
-            .status(HttpStatus.BAD_REQUEST)
-            .body(ErrorResponse(e.message))
-    }
-
-}
