--- src/main/kotlin/com/sparta/reviewus/auth/exception/ExpiredJwtException.kt
@@ -1,6 +1,0 @@
-package com.sparta.reviewus.auth.exception
-
-import io.jsonwebtoken.JwtException
-
-data class ExpiredJwtException(override val message: String?) : JwtException(message) {
-}
