--- src/main/kotlin/com/sparta/reviewus/common/Base64Util.kt
@@ -1,19 +1,0 @@
-package com.sparta.reviewus.common
-
-import java.util.*
-
-object Base64Util {
-    fun decodeBasicAuth(headerValue: String): Array<String> {
-        getEncodedAuthFromAuthorizationHeader(headerValue).let {
-            Base64.getDecoder().decode(it)
-        }.let {
-            String(it).split(":")
-        }.let {
-            return arrayOf(it[0], it[1])
-        }
-    }
-
-    private fun getEncodedAuthFromAuthorizationHeader(headerValue: String): String {
-        return headerValue.split(" ")[1]
-    }
-}
