+++ src/main/kotlin/com/sparta/reviewus/domain/exception/GlobalExceptionHandler.kt
@@ -1,6 +1,7 @@
 package com.sparta.reviewus.domain.exception
 
 import com.sparta.reviewus.domain.exception.dto.ErrorResponse
+import com.sparta.reviewus.domain.exception.member.BadCredentialsException
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.MethodArgumentNotValidException
@@ -27,6 +28,13 @@ class GlobalExceptionHandler {
     fun handleModelNotFoundException(e: ModelNotFoundException): ResponseEntity<ErrorResponse> {
         return ResponseEntity
             .status(HttpStatus.NOT_FOUND)
+            .body(ErrorResponse(e.message))
+    }
+
+    @ExceptionHandler(BadCredentialsException::class)
+    fun handleModelNotFoundException(e: BadCredentialsException): ResponseEntity<ErrorResponse> {
+        return ResponseEntity
+            .status(HttpStatus.BAD_REQUEST)
             .body(ErrorResponse(e.message))
     }
