+++ src/main/kotlin/com/sparta/reviewus/domain/member/repository/MemberRepository.kt
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 
 interface MemberRepository: JpaRepository<Member, Long> {
     fun existsMemberByEmail(email: String): Boolean
-    fun findMemberByEmail(email: String): Member?
     fun findMemberByEmailAndPassword(email: String, password: String): Member?
 }
