+++ src/main/kotlin/com/sparta/reviewus/domain/member/service/MemberService.kt
@@ -1,7 +1,6 @@
 package com.sparta.reviewus.domain.member.service
 
 import com.sparta.reviewus.domain.member.dto.JoinRequest
-import com.sparta.reviewus.domain.member.dto.LoginRequest
 import com.sparta.reviewus.domain.member.dto.MemberResponse
 import com.sparta.reviewus.domain.member.dto.ProfileUpdateRequest
 
@@ -14,7 +13,5 @@ interface MemberService {
     fun updateMemberProfile(profileUpdateRequest: ProfileUpdateRequest): MemberResponse
 
     fun join(joinRequest: JoinRequest): String
-
-    fun login(loginRequest: LoginRequest): String
 
 }
