+++ src/main/kotlin/com/sparta/reviewus/domain/member/service/MemberServiceImpl.kt
@@ -1,11 +1,7 @@
 package com.sparta.reviewus.domain.member.service
 
-import com.sparta.reviewus.common.JWTUtil
-import com.sparta.reviewus.domain.exception.DuplicatedPropertyException
 import com.sparta.reviewus.domain.exception.ModelNotFoundException
-import com.sparta.reviewus.auth.exception.BadCredentialsException
 import com.sparta.reviewus.domain.member.dto.JoinRequest
-import com.sparta.reviewus.domain.member.dto.LoginRequest
 import com.sparta.reviewus.domain.member.dto.MemberResponse
 import com.sparta.reviewus.domain.member.dto.ProfileUpdateRequest
 import com.sparta.reviewus.domain.member.model.Member
@@ -25,20 +21,13 @@ class MemberServiceImpl(
         val (email, password, name, nickname) = joinRequest
 
         //이메일 중복 체크
-        if (memberRepository.existsMemberByEmail(email)) throw DuplicatedPropertyException(
+        if (memberRepository.existsMemberByEmail(email)) throw com.sparta.reviewus.domain.exception.DuplicatedPropertyException(
             "이메일",
             email
         )
 
         memberRepository.save(Member(email = email, password = password, name = name, profile = Profile(nickname = nickname)))
         return "회원가입 성공"
-    }
-
-    override fun login(loginRequest: LoginRequest): String {
-        val authenticatedMember =
-            memberRepository.findMemberByEmailAndPassword(loginRequest.email, loginRequest.password) ?: throw BadCredentialsException()
-
-        return JWTUtil.generateToken(authenticatedMember.email)
     }
 
     override fun getMemberProfile(): MemberResponse {
